# refactor(swc/plugin): revise module cache implementation