--- a/crates/swc_plugin_runner/Cargo.toml
+++ b/crates/swc_plugin_runner/Cargo.toml
@@ -16,7 +16,7 @@ parking_lot = "0.12.0"
 serde = {version = "1.0.126", features = ["derive"]}
 serde_json = "1.0.64"
 swc_atoms = {version = "0.2.7", path = '../swc_atoms'}
-swc_common = {version = "0.17.0", path = "../swc_common", features = ["plugin-rt"]}
+swc_common = {version = "0.17.0", path = "../swc_common", features = ["plugin-rt", "concurrent"]}
 swc_ecma_ast = {version = "0.65.0", path = "../swc_ecma_ast", features = ["rkyv-impl"]}
 swc_ecma_loader = { version = "0.28.0", path = "../swc_ecma_loader" }
 swc_ecma_parser = {version = "0.88.0", path = "../swc_ecma_parser"}