… crates/swc_plugin_runner/src/resolve.rs
… tests/rust-plugins/swc_internal_plugin/Cargo.lock