<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

This PR attempts to revise our plugin module cache. 

It looks like the current implementation is obviously flawed spends huge amount of time to initialize module. PR moves out cache into shared, static instance across, as well as more aggressively utilize in-memory cache for module instance.

- Cache compiled `Module` instance in-memory instead of bytecodes
Previously we keep compiled bytecode of plugin (`vec<u8>`) in its memory then create module instance from it. It doesn't necessarily have to be. Now we're keeping actual instance of `Module` in memory, instance creation occurs only once when we load cold bytes (either from `fs` or `fs_cache` bytecode)

- bytecode cache versioning
Pretty basic, but introduces versioning for cache root path for compiled byte cache stability. It's fairly safe to move bytecache around machine but it is not safe against different version of `wasmer`. It's bit manual process we need to update version string though.


**Testing**

I used below snippet (https://github.com/swc-project/swc/issues/3598#issuecomment-1042416762) against our `swc_internal_plugin` to see number differences.

```ts
const EXECUTE_COUNT = 10000;

...

const arr = [...Array(EXECUTE_COUNT).keys()];
obs.observe({ entryTypes: ['measure'], buffered: true });
let tasks = arr.map((value) => transform(`console.log(${value})`, {
  jsc: {
    experimental: {
      plugins: [[`${pluginPath}`, {}]]
    }
  }
}));

performance.mark("executeTask-start");
Promise.all(tasks).then(() => {
  performance.mark("executeTask-complete");
  performance.measure("executeTask", "executeTask-start", "executeTask-complete");
});
```

Which schedules 10,000 transforms in async with plugin init phase. Number may vary across devices.

**Before**
```
PerformanceEntry {
  name: 'executeTask',
  entryType: 'measure',
  startTime: 0,
  duration: 46487.8278
}
```

**after**
```
❯❯❯ hyperfine --warmup "2" "node test.js"
Benchmark 1: node test.js
  Time (mean ± σ):      1.335 s ±  0.046 s    [User: 0.003 s, System: 0.009 s]
  Range (min … max):    1.277 s …  1.434 s    10 runs

❯❯❯ node test.js

PerformanceEntry {
  name: 'executeTask',
  entryType: 'measure',
  startTime: 0,
  duration: 1429.4356
}
```

Which shows clear differences between two.

**Related issue (if exists):**

- https://github.com/swc-project/swc/issues/3598